fix: address codex token stats review
